+++ main.rb
@@ -1,8 +1,16 @@
 require_relative 'app'
+require_relative 'main_fun'
+require_relative 'handle_choice'
 
-def main
-  app = App.new
-  app.run
+def show_menu
+  puts 'Please select an option:'
+  puts '1. List all books'
+  puts '2. List all people'
+  puts '3. Create a person'
+  puts '4. Create a book'
+  puts '5. Create a rental'
+  puts '6. List rentals for a person'
+  puts '7. Quit'
 end
 
 main
